Web App/requirements.txt
# --find-links https://download.pytorch.org/whl/torch_stable.html
# torch==2.3.0
# torchvision==0.18.0
pillow==10.3.0
pandas==2.2.1
gevent==23.9.1
flask==3.0.3
flask_cors==4.0.2
openpyxl==3.1.2
